+++ routes/products.js
@@ -2,30 +2,39 @@ const express = require("express");
 const router = express.Router();
 const Joi = require("joi");
 
-const products = [
-    { id: 1, name: "iphone 12", price: 20000 },
-    { id: 2, name: "iphone 13", price: 30000 },
-    { id: 3, name: "iphone 14", price: 40000 }
-];
+const {Product,validateProduct} = require("../models/product");
+const { describe } = require("node:test");
+const { join } = require("path");
 
-router.get("/", (req, res) => {
+router.get("/", async (req, res) => {
+    const products = await Product.find();
     res.send(products);
 });
 
-router.post("/", (req, res) => {
+router.post("/", async (req, res) => {
     const { error } =  validateProduct(req.body);
 
     if(error) {
-        return res.status(400).send(result.error.details[0].message);
+        return res.status(400).send(error.details[0].message);
     }
 
-    const product = {
-        id: products.length + 1,
+    const product = new Product({
         name: req.body.name,
-        price: req.body.price
-    };
-    products.push(product);
-    res.send(product);
+        price: req.body.price,
+        description: req.body.description,
+        imageUrl: req.body.imageUrl,
+        isActive: req.body.isActive
+    });
+
+    try{
+        const result = await product.save();
+        res.send(result);
+    }
+
+    catch(err){
+        console.log(err);
+    }
+
 });
 
 router.put("/:id", (req, res) => {
@@ -66,13 +75,6 @@ router.get("/:id", (req, res) => {
     res.send(product);
 });
 
-function validateProduct(product) {
-    const schema = new Joi.object({
-        name: Joi.string().min(3).max(30).required(),
-        price: Joi.number().required()
-    });
 
-    return schema.validate(product);
-}
 
 module.exports = router;
